Updated plot function!!

+++ AddressBook.py
@@ -335,7 +335,7 @@ class AddressBook:
             wallet_vector.valueUSD = np.multiply(wallet_vector.valueUSD, wallet_vector.tx_type)
         plt.gca().xaxis.set_major_formatter(mdates.DateFormatter('%Y-%m-%d'))
         plt.gca().xaxis.set_major_locator(mdates.DayLocator())
-        if not wallet_vector.time.dtype == int:
+        if not wallet_vector.time.dtype == float:
             return
         scatter = plt.scatter(wallet_vector.time.apply(datetime.datetime.fromtimestamp),
                               wallet_vector.valueUSD,
